Remove dependency on openai client, smart defaults
I'm removing the dependency on the openai client, we only use it for a
single openai endpoint. It makes configuration a little weird because
not all configurations can be customized at runtime. I don't want you to
have to configure openai client globally incase you're using it
elsewhere with different constraints. This allows me to change the
defaults to enable streaming which leads to a much simpler install
experience.

This also updates the tests and the adapter protocol to properly plumb
the config through the instructor library as it should've been.